requirements.txt
pyTelegramBotAPI==4.14.0
google-generativeai>=0.5.0
python-dotenv==1.0.0
pillow
python-pptx
requests
lxml
duckduckgo-search
